[scenes list] towards editable

+++ app/lib/src/projects/project_list.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:robokru/src/data/tables.dart';
 import 'package:robokru/src/projects/selected_project.dart';
@@ -7,9 +6,7 @@ import 'package:robokru/src/skeleton/consumer_async_value_widget.dart';
 
 import '../data/query/projects.dart';
 import '../data/id.dart';
-import '../settings/settings_view.dart';
 import '../skeleton/top_level_navigation_drawer.dart';
-import 'project_scene_list.dart';
 
 final projectsProvider = StreamProvider.autoDispose((ref) {
   final skaidb = ref.watch(projectsDaoProvider);
@@ -52,8 +49,36 @@ class ProjectListView extends ConsumerWidget {
                   trailing: IconButton(
                     icon: const Icon(Icons.delete),
                     onPressed: () async {
-                      final skaidb = ref.read(projectsDaoProvider);
-                      await skaidb.deleteProject(item.id);
+                      // ask for confirmation
+                      final bool? delete = await showDialog<bool>(
+                        context: context,
+                        builder: (BuildContext context) {
+                          return AlertDialog(
+                            title: const Text('Delete Project?'),
+                            content: Text(
+                                'Are you sure you want to delete ${item.name}?'),
+                            actions: <Widget>[
+                              TextButton(
+                                child: const Text('Cancel'),
+                                onPressed: () {
+                                  Navigator.of(context).pop(false);
+                                },
+                              ),
+                              TextButton(
+                                child: const Text('Delete'),
+                                onPressed: () {
+                                  Navigator.of(context).pop(true);
+                                },
+                              ),
+                            ],
+                          );
+                        },
+                      );
+
+                      if (delete == true) {
+                        final skaidb = ref.read(projectsDaoProvider);
+                        await skaidb.deleteProject(item.id);
+                      }
                     },
                   ),
                   onTap: () {
@@ -77,14 +102,6 @@ class ProjectListView extends ConsumerWidget {
         return Scaffold(
           appBar: AppBar(
             title: const Text('Projects'),
-            actions: [
-              IconButton(
-                icon: const Icon(Icons.settings),
-                onPressed: () {
-                  context.pushNamed(SettingsView.routeName);
-                },
-              ),
-            ],
           ),
 
           // To work with lists that may contain a large number of items, it’s best
